Make bulk code transfers to FileProcessor and FileUtils
